requirements.txt
pandas
geopandas
shapely
plotly
requests
pyarrow
pytest
kaleido
Pillow
